requirements.txt
fastapi>=0.127.0
uvicorn>=0.39.0
rembg>=2.0.0
Pillow>=10.0.0
mediapipe>=0.10.0
numpy>=1.24.0
python-multipart>=0.0.6
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0
sqlalchemy>=2.0.0
git+https://github.com/openai/CLIP.git@main
